back-end/requirements.txt
Django==4.0.6
django-cors-headers==3.13.0
django_filter==22.1
djangorestframework==3.13.1
matplotlib==3.5.2
numpy==1.23.1
pandas==1.4.3
python-dotenv==0.20.0
requests==2.28.1
scipy==1.8.1
psycopg2==2.9.3
chess==1.9.2